+++ compilers/concrete-compiler/compiler/Makefile
@@ -150,6 +150,8 @@ $(BUILD_DIR)/configured.stamp:
 	-DLLVM_TARGETS_TO_BUILD="host" \
 	-DCMAKE_BUILD_TYPE=$(BUILD_TYPE) \
 	-DLLVM_ENABLE_ASSERTIONS=ON \
+	-DLLVM_ENABLE_RTTI=ON \
+	-DLLVM_LINK_LLVM_DYLIB=on \
 	-DMLIR_ENABLE_BINDINGS_PYTHON=$(BINDINGS_PYTHON_ENABLED) \
 	-DCONCRETELANG_BINDINGS_PYTHON_ENABLED=$(BINDINGS_PYTHON_ENABLED) \
 	-DCONCRETELANG_DATAFLOW_EXECUTION_ENABLED=$(DATAFLOW_EXECUTION_ENABLED) \
